ref: replace first/last with generic aggregators
This allows us to eliminate the "pushdown" code for applying first/last
to records, which were exceedingly complicated.
