+++ crates/sparrow-instructions/src/evaluators/aggregation.rs
@@ -1,21 +1,17 @@
-mod boolean;
+mod first_evaluator;
 mod function;
 mod generic;
-mod list;
-mod map;
+mod last_evaluator;
 mod numeric_properties;
 mod primitive;
-mod string;
 mod token;
 mod two_stacks;
 
-pub use boolean::*;
+pub use first_evaluator::*;
 pub use function::*;
 pub use generic::*;
-pub use list::*;
-pub use map::*;
+pub use last_evaluator::*;
 pub use numeric_properties::*;
 pub use primitive::*;
-pub use string::*;
 pub use token::*;
 pub use two_stacks::*;
