--- crates/sparrow-instructions/src/evaluators/aggregation/boolean.rs
@@ -1,10 +1,0 @@
-//! Boolean aggregation evaluators.
-
-mod first_boolean_evaluator;
-mod last_boolean_evaluator;
-
-mod two_stacks_first_boolean_evaluator;
-mod two_stacks_last_boolean_evaluator;
-
-pub use first_boolean_evaluator::*;
-pub use last_boolean_evaluator::*;
