+++ crates/sparrow-instructions/src/evaluators/aggregation/function.rs
@@ -1,13 +1,9 @@
 //! Aggregation functions.
 
 pub mod agg_fn;
-pub mod boolean_agg_fn;
 pub mod count_agg_fn;
 pub mod primitive_agg_fn;
-pub mod string_agg_fn;
 
 pub use agg_fn::*;
-pub use boolean_agg_fn::*;
 pub use count_agg_fn::*;
 pub use primitive_agg_fn::*;
-pub use string_agg_fn::*;
