+++ crates/sparrow-instructions/src/evaluators/aggregation/function/primitive_agg_fn.rs
@@ -70,101 +70,6 @@ where
     }
 }
 
-/// Placeholder struct for the implementation of the [[AggFn]] for
-/// `first` aggregation on primitives.
-pub struct FirstPrimitive<T: ArrowPrimitiveType>
-where
-    T::Native: Copy,
-{
-    // Make the compiler happy by using the type parameter. Conceptually, the aggregation
-    // will store values of type T.
-    _phantom: PhantomData<T>,
-}
-
-impl<T: ArrowPrimitiveType> ArrowAggFn for FirstPrimitive<T> {
-    type InArrowT = T;
-    type OutArrowT = T;
-}
-
-impl<T: ArrowPrimitiveType> AggFn for FirstPrimitive<T> {
-    type InT = T::Native;
-    type AccT = Option<T::Native>;
-    type OutT = T::Native;
-
-    fn zero() -> Self::AccT {
-        None
-    }
-
-    fn one(input: &Self::InT) -> Self::AccT {
-        Some(*input)
-    }
-
-    fn merge(acc1: &mut Self::AccT, acc2: &Self::AccT) {
-        if acc2.is_some() && acc1.is_none() {
-            *acc1 = *acc2;
-        }
-    }
-
-    fn extract(acc: &Self::AccT) -> Option<Self::OutT> {
-        *acc
-    }
-
-    fn add_one(acc: &mut Self::AccT, input: &Self::InT) {
-        if acc.is_none() {
-            *acc = Some(*input)
-        }
-    }
-
-    fn name() -> &'static str {
-        "first"
-    }
-}
-
-/// Placeholder struct for the implementation of the [[AggFn]] for
-/// `last` aggregation on primitives.
-pub struct LastPrimitive<T: ArrowPrimitiveType> {
-    // Make the compiler happy by using the type parameter. Conceptually, the aggregation
-    // will store values of type T.
-    _phantom: PhantomData<T>,
-}
-
-impl<T: ArrowPrimitiveType> ArrowAggFn for LastPrimitive<T> {
-    type InArrowT = T;
-    type OutArrowT = T;
-}
-
-impl<T: ArrowPrimitiveType> AggFn for LastPrimitive<T> {
-    type InT = T::Native;
-    type AccT = Option<T::Native>;
-    type OutT = T::Native;
-
-    fn zero() -> Self::AccT {
-        None
-    }
-
-    fn one(input: &Self::InT) -> Self::AccT {
-        Some(*input)
-    }
-
-    fn merge(acc1: &mut Self::AccT, acc2: &Self::AccT) {
-        if acc2.is_some() {
-            *acc1 = *acc2
-        }
-    }
-
-    fn extract(acc: &Self::AccT) -> Option<Self::OutT> {
-        *acc
-    }
-
-    fn add_one(acc: &mut Self::AccT, input: &Self::InT) {
-        *acc = Some(*input)
-    }
-
-    fn name() -> &'static str {
-        "last"
-    }
-}
-
 /// Placeholder struct for the implementation of the [[AggFn]] for `max`
 /// aggregation.
 pub struct Max<T: ArrowPrimitiveType>
