--- crates/sparrow-instructions/src/evaluators/aggregation/list.rs
@@ -1,7 +1,0 @@
-//! List aggregation evaluators.
-
-mod first_list_evaluator;
-mod last_list_evaluator;
-
-pub use first_list_evaluator::*;
-pub use last_list_evaluator::*;
