--- crates/sparrow-instructions/src/evaluators/aggregation/map.rs
@@ -1,7 +1,0 @@
-//! Map aggregation evaluators.
-
-mod first_map_evaluator;
-mod last_map_evaluator;
-
-pub use first_map_evaluator::*;
-pub use last_map_evaluator::*;
