--- crates/sparrow-instructions/src/evaluators/aggregation/string.rs
@@ -1,10 +1,0 @@
-//! String aggregation evaluators.
-
-mod first_string_evaluator;
-mod last_string_evaluator;
-
-mod two_stacks_first_string_evaluator;
-mod two_stacks_last_string_evaluator;
-
-pub use first_string_evaluator::*;
-pub use last_string_evaluator::*;
