+++ crates/sparrow-instructions/src/evaluators/aggregation/token.rs
@@ -1,27 +1,17 @@
 //! Tokens representing keys for compute storage.
 
-mod boolean_accum_token;
+mod array_ref_accum_token;
 mod collect_struct_token;
 mod collect_token;
 mod count_accum_token;
-mod list_accum_token;
-mod map_accum_token;
 mod primitive_accum_token;
-mod string_accum_token;
-mod two_stacks_boolean_accum_token;
 mod two_stacks_count_accum_token;
 mod two_stacks_primitive_accum_token;
-mod two_stacks_string_accum_token;
 
-pub use boolean_accum_token::*;
+pub use array_ref_accum_token::*;
 pub use collect_struct_token::*;
 pub use collect_token::*;
 pub use count_accum_token::*;
-pub use list_accum_token::*;
-pub use map_accum_token::*;
 pub use primitive_accum_token::*;
-pub use string_accum_token::*;
-pub use two_stacks_boolean_accum_token::*;
 pub use two_stacks_count_accum_token::*;
 pub use two_stacks_primitive_accum_token::*;
-pub use two_stacks_string_accum_token::*;
